Assignment7-final

+++ NavCtrl/Company.h
@@ -1,7 +1,7 @@
 //
 //  Company.h
 //  NavCtrl
-// Assignment7-MMM
+// Assignment7
 // Manual Memory Management
 //
 //  Created by Emiko Clark on 2/29/16.
@@ -14,12 +14,15 @@
 @interface Company : NSObject
 
 @property int companyID;
-@property int row;
+@property float row;
 @property (nonatomic, strong) NSString *name;
 @property (nonatomic, strong) NSString *stockSymbol;
 @property (nonatomic, strong) NSString *stockPrice;
 @property (nonatomic, strong) NSString *logo;
 @property (nonatomic, strong) NSMutableArray <Product*>  *productArray;
 
+
+
+//- (id) initWithName:(NSString*)name andStockSymbol:(NSString *)stockSymbol andLogo:(NSString*)logo;
 - (id) initWithName:(NSString*)name andStockSymbol:(NSString *)stockSymbol andLogo:(NSString*)logo andRow:(float)row andCompanyID:(int)companyID;
 @end
