+++ NavCtrl/Product.m
@@ -1,17 +1,18 @@
 //
 //  Product.m
 //  NavCtrl
-// Assignment7-MMM
+// Assignment7
 // Manual Memory Management
 //
 //  Created by Emiko Clark on 2/29/16.
 //  Copyright © 2016 Aditya Narayan. All rights reserved.
+//
 
 #import "Product.h"
 
 @implementation Product
 
-- (id) initWithName:(NSString*)name andUrl:(NSString*)url andLogo:(NSString*)logo andCompanyID:(NSInteger)companyID andRow:(float)row {
+- (id) initWithName:(NSString*)name andUrl:(NSString*)url andLogo:(NSString*)logo andCompanyID:(int)companyID andRow:(float)row {
     self=[super init];
     if (self) {
         self.name = name;
@@ -24,6 +25,6 @@
 }
 
 -(NSString*)description {
-    return [NSString stringWithFormat:@"%@, CC:%ld, CP:%ld, row:%f, %@, %@", self.name, self.companyID, self.productID, self.row, self.url, self.logo];
+    return [NSString stringWithFormat:@"%@, CC:%i, CP:%i, row:%f, %@, %@", self.name, self.companyID, self.productID, self.row, self.url, self.logo];
 }
 @end
